Multi-Step Reasoning Agent/requirements.txt
google-generativeai==0.8.3
pytest==7.4.0
flask==3.0.0
flask-cors==4.0.0

